+++ query_record_command.go
@@ -116,7 +116,7 @@ func (cmd *queryRecordCommand) parseRecordResults(ifc command, receiveSize int) 
 		for {
 			select {
 			// send back the result on the async channel
-			case cmd.recordset.Records <- newRecord(cmd.node, key, bins, nil, generation, expiration):
+			case cmd.recordset.Records <- newRecord(cmd.node, key, bins, generation, expiration):
 				break L
 			case <-cmd.recordset.cancelled:
 				return false, NewAerospikeError(SCAN_TERMINATED)