+++ read_header_command.go
@@ -59,7 +59,7 @@ func (cmd *readHeaderCommand) parseResult(ifc command, conn *Connection) error {
 	if resultCode == 0 {
 		generation := int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 14)))
 		expiration := TTL(int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 18))))
-		cmd.record = newRecord(cmd.node, cmd.key, nil, nil, generation, expiration)
+		cmd.record = newRecord(cmd.node, cmd.key, nil, generation, expiration)
 	} else {
 		if ResultCode(resultCode) == KEY_NOT_FOUND_ERROR {
 			cmd.record = nil